requirements.txt
scapy
requests
netifaces
paramiko
python-nmap
ipaddress